pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fakenews-tribunal"
version = "0.1.0"
description = "Autonomous multi-agent fact-checking system"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "uvicorn[standard]>=0.41.0",
    "litellm>=1.81.11",
    "tavily-python>=0.7.21",
    "sqlalchemy[asyncio]>=2.1.0",
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "python-jose[cryptography]>=3.5.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.13.1",
    "structlog>=25.5.0",
    "typer>=0.24.1",
    "httpx>=0.28.1",
]

[project.scripts]
tribunal = "cli.main:app"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.9.0",
    "mypy>=1.14.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["B008"]

[tool.mypy]
python_version = "3.12"
strict = false
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["api", "agents", "llm", "tools", "db", "core", "cli"]
